+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Bugfixes:
 * Fix compatibility with Fig (Terminal, macOS)
 * Fix option `--title-fqdn` doesn't work (Title)
 * Fix used spaces calculation (Disk, Linux / BSD / macOS, #508)
+* Fix `--brightness-format` (Brightness)
 
 Logo:
 * Change the special handling of `kitty` protocol with `.png` image file to a new image protocol `kitty-direct`. This is the fastest image protocol because fastfetch doesn't need to pre-encode the image to base64 or something and the image content doesn't need to be transmitted via tty. Note: